# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KasiDash & BuildForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders with optional status filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, fulfilled, cancelled, refunded]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersPage"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/fulfil:
    post:
      operationId: fulfilOrder
      tags: [orders]
      summary: Mark order as fulfilled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, draft, archived]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/bulk-update:
    post:
      operationId: bulkUpdateProducts
      tags: [products]
      summary: Bulk update product prices/status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkProductUpdate"
      responses:
        "200":
          description: Updated count
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                required: [updated]

  # Customers
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: low_stock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryPage"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update stock quantity for an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [pending, paid, fulfilled, cancelled, refunded]
        total:
          type: number
        itemCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, customerName, customerEmail, status, total, itemCount, createdAt, updatedAt]
    OrdersPage:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [orders, total, page, limit]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, draft, archived]
        inventory:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, price, status, inventory, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, draft, archived]
        inventory:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
      required: [title, price, status, inventory]
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, draft, archived]
        inventory:
          type: integer
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
    ProductsPage:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]
    BulkProductUpdate:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        price:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
      required: [ids]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        sku:
          type: ["string", "null"]
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, productId, productTitle, quantity, lowStockThreshold, isLowStock, updatedAt]
    InventoryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
    InventoryUpdate:
      type: object
      properties:
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
      required: [quantity]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_order, low_inventory, failed_payment, high_sales, app_update]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, message, isRead, createdAt]
...