# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * KasiDash & BuildForge API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from './productInputStatus';

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  status: ProductInputStatus;
  inventory: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
}